feat: update EnabledJobs and Relay modal
